TodoApp add projects

+++ src/main/java/com/example/todoapp/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package com.example.todoapp.repository;
 
-import com.example.todoapp.entity.User;
+import com.example.todoapp.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByEmail(String email);
